perf(treesitter): make treesitter load faster
Add nvim-treesitter queries to the rtp and it's custom query predicates early
This is needed because a bunch of plugins no longer `require("nvim-treesitter")`, which
no longer trigger the **nvim-treesitter** module to be loaded in time.
Luckily, the only thins that those plugins need are the custom queries, which we make available
during startup.